+++ app/database/database.py
@@ -9,6 +9,10 @@ SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 Base = declarative_base()
 
 
+def initialize_db():
+    Base.metadata.create_all(bind=engine)
+
+
 def get_db():
     db: Session = SessionLocal()
     try:
