+++ app/schemas/group.py
@@ -15,11 +15,11 @@ class Group(GroupBase):
     users: List[str]
 
     class Config:
-        orm_mode = True
+        from_attributes = True
 
 
 class GroupUpdate(BaseModel):
     name: Optional[str] = None
 
     class Config:
-        orm_mode = True
+        from_attributes = True
